requirements.txt
alembic==1.14.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.12
Flask==3.1.0
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.38
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
